Split index to cli
